requirements.txt
streamlit>=1.30.0
google-genai>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
keyring>=25.0.0
